[Chore] #515 - Notification.Name을 StringLiterals에서 별개의 enum으로 이동
