+++ WSSiOS/WSSiOS/Resource/Constants/Strings/StringLiterals.swift
@@ -12,21 +12,6 @@ enum StringLiterals {
         static let deviceIdentifier = "DEVICE_IDENTIFIER"
     }
     
-    enum NotificationCenter {
-        static let novelReviewKeywordSelected = Notification.Name("novelReviewKeywordSelected")
-        static let novelReviewDataSelected = Notification.Name("novelReviewDataSelected")
-        static let novelReviewDateRemoved = Notification.Name("novelReviewDateRemoved")
-        static let feedEdited = Notification.Name("feedEdited")
-        static let novelReviewed = Notification.Name("novelReviewed")
-        static let popFeedDetailViewController = Notification.Name("popFeedDetailViewController")
-        static let feedNovelConnected = Notification.Name("feedNovelConnected")
-        static let blockUser = Notification.Name("blockUser")
-        static let pushToUpdateDetailSearchResult = Notification.Name("pushToUpdateDetailSearchResult")
-        static let pushToDetailSearchResult = Notification.Name("pushToDetailSearchResult")
-        static let changeRepresentativeAvatar = Notification.Name("changeRepresentativeAvatar")
-        static let moveToLibraryTab = Notification.Name("moveToLibraryTab")
-    }
-    
     enum UserDefault {
         static let accessToken = "ACCESS_TOKEN"
         static let refreshToken = "REFRESH_TOKEN"
